+++ src/captioner.rs
@@ -1,15 +1,18 @@
 use std::io::Write;
+use std::str::FromStr;
 use std::{fs, io::Read, path::PathBuf};
 
 use anyhow::anyhow;
-use base64::{engine::general_purpose, Engine as _};
+use aws_sdk_bedrockruntime::primitives::Blob;
+use aws_sdk_bedrockruntime::types::{
+    ContentBlock, ImageBlock, ImageFormat, ImageSource, InferenceConfiguration, SystemContentBlock,
+};
 
 use indicatif::{ProgressBar, ProgressStyle};
 use quick_xml::se;
 use serde::Serialize;
 
-use crate::ask_bedrock;
-use crate::RunType;
+use crate::models::converse::call_converse;
 
 #[derive(Debug, Serialize)]
 pub struct Image {
@@ -18,7 +21,8 @@ pub struct Image {
     pub extension: String,
     // FIX: Think about setting the base64 as optional
     #[serde(skip_serializing)]
-    pub base64: String,
+    //pub base64: String,
+    pub base64: Vec<u8>,
     pub caption: Option<String>,
 }
 
@@ -66,22 +70,34 @@ pub fn list_files_in_path_by_extension(
     Ok(files)
 }
 
-pub fn load_image(p: &PathBuf) -> Result<String, anyhow::Error> {
+//pub fn load_image(p: &PathBuf) -> Result<String, anyhow::Error> {
+pub fn load_image(p: &PathBuf) -> Result<Vec<u8>, anyhow::Error> {
     let mut file = fs::File::open(p)?;
     let mut buffer = Vec::new();
     file.read_to_end(&mut buffer)?;
 
-    let base64_img = general_purpose::STANDARD.encode(buffer);
-    Ok(base64_img)
+    //let base64_img = general_purpose::STANDARD.encode(buffer);
+    //Ok(base64_img)
+    Ok(buffer)
 }
 
 pub async fn caption_image(
     i: &mut Vec<crate::captioner::Image>,
     model: &str,
-    prompt: &String,
+    prompt: &str,
     runtime_client: &aws_sdk_bedrockruntime::Client,
-    bedrock_client: &aws_sdk_bedrock::Client,
+    _bedrock_client: &aws_sdk_bedrock::Client,
 ) -> Result<(), anyhow::Error> {
+    // Needs to be hardcoded for images
+    let inference_parameters: InferenceConfiguration = InferenceConfiguration::builder()
+        .max_tokens(2048)
+        .top_p(0.8)
+        .temperature(0.5)
+        .build();
+
+    // FIX: Remove the clone
+    let system_prompt = Some(vec![SystemContentBlock::Text(prompt.to_owned())]);
+
     // progress bar shenanigans
     let progress_bar = ProgressBar::new(i.len().try_into()?);
     progress_bar.set_style(
@@ -95,14 +111,37 @@ pub async fn caption_image(
 
     for image in i {
         let message = image.path.as_path().display().to_string();
+
+        //let imagesrc: ImageSource = ImageSource::Bytes(image.base64.to_blob());
+        let imagesrc: ImageSource = ImageSource::Bytes(
+            // FIX: Try to remove the clone
+            Blob::new(image.base64.clone()),
+        );
+
+        let image_block = ImageBlock::builder()
+            .source(imagesrc)
+            .format(ImageFormat::from_str(image.extension.as_str())?)
+            .build()?;
+
+        let content = ContentBlock::Image(image_block);
         progress_bar.set_message(message);
-        let caption = ask_bedrock(
-            prompt,
-            Some(image),
-            model,
-            RunType::Captioning,
+        // let caption = ask_bedrock(
+        //     prompt,
+        //     Some(image),
+        //     model,
+        //     RunType::Captioning,
+        //     runtime_client,
+        //     bedrock_client,
+        // )
+        // .await?;
+        let caption = call_converse(
             runtime_client,
-            bedrock_client,
+            model.to_string(),
+            // FIX: Avoid the clone
+            inference_parameters.clone(),
+            content,
+            // FIX: Avoid the clone
+            system_prompt.clone(),
         )
         .await?;
         progress_bar.inc(1);
@@ -143,6 +182,7 @@ pub fn write_captions(
 mod tests {
     use super::*;
     use crate::utils::load_bedrust_config;
+    use base64::{engine::general_purpose, Engine as _};
     use image::{Rgb, RgbImage};
     use rand::distributions::{Alphanumeric, DistString};
 
@@ -182,11 +222,10 @@ mod tests {
         ];
         assert_eq!(expected_vec, list.unwrap());
     }
-
     #[test]
     fn load_image_from_disk() {
         let image_path = PathBuf::from("/tmp/bedrust_test_image.jpeg");
-        // generate an image
+        // Generate an image
         let mut img = RgbImage::new(32, 32);
         for x in 15..=17 {
             for y in 8..24 {
@@ -196,11 +235,18 @@ mod tests {
         }
         img.save(&image_path).unwrap();
 
-        // load the generated image from disk
+        // Load the generated image from disk
         let test_image = load_image(&PathBuf::from(&image_path)).unwrap();
 
-        // this is just raw base64 of the generated image from above
-        let image_base64 = "/9j/4AAQSkZJRgABAgAAAQABAAD/wAARCAAgACADAREAAhEBAxEB/9sAQwAIBgYHBgUIBwcHCQkICgwUDQwLCwwZEhMPFB0aHx4dGhwcICQuJyAiLCMcHCg3KSwwMTQ0NB8nOT04MjwuMzQy/9sAQwEJCQkMCwwYDQ0YMiEcITIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIy/8QAHwAAAQUBAQEBAQEAAAAAAAAAAAECAwQFBgcICQoL/8QAtRAAAgEDAwIEAwUFBAQAAAF9AQIDAAQRBRIhMUEGE1FhByJxFDKBkaEII0KxwRVS0fAkM2JyggkKFhcYGRolJicoKSo0NTY3ODk6Q0RFRkdISUpTVFVWV1hZWmNkZWZnaGlqc3R1dnd4eXqDhIWGh4iJipKTlJWWl5iZmqKjpKWmp6ipqrKztLW2t7i5usLDxMXGx8jJytLT1NXW19jZ2uHi4+Tl5ufo6erx8vP09fb3+Pn6/8QAHwEAAwEBAQEBAQEBAQAAAAAAAAECAwQFBgcICQoL/8QAtREAAgECBAQDBAcFBAQAAQJ3AAECAxEEBSExBhJBUQdhcRMiMoEIFEKRobHBCSMzUvAVYnLRChYkNOEl8RcYGRomJygpKjU2Nzg5OkNERUZHSElKU1RVVldYWVpjZGVmZ2hpanN0dXZ3eHl6goOEhYaHiImKkpOUlZaXmJmaoqOkpaanqKmqsrO0tba3uLm6wsPExcbHyMnK0tPU1dbX2Nna4uPk5ebn6Onq8vP09fb3+Pn6/9oADAMBAAIRAxEAPwD5/oAKACgAoAKANKwsLxNRtWa0nVVmQkmMgAZHtXNWrU3TklJbPqe1l2XYyGMpSlSkkpR+y+68ixrlldzazcPHazOh24ZYyQflFZ4SrTjRScl9/mdvEGAxVXMak6dKTTtqk2vhXkYtdp8wFAG1Za5qM1/bxvcZR5VVhsXkE/SuKrhKMacml0fc+nwHEGY1cVSpzqXTkk9I7NryJ9Y1i/tdVmhhn2xrtwNin+EHuKzw2GpTpKUlr8+51Z3nePw2PqUaNS0VaysuyfVHPV6J8cFABQAUAFAH/9k=";
-        assert_eq!(test_image, image_base64);
+        // This is just raw base64 of the generated image from above
+        let expected_image_base64 = "/9j/4AAQSkZJRgABAgAAAQABAAD/wAARCAAgACADAREAAhEBAxEB/9sAQwAIBgYHBgUIBwcHCQkICgwUDQwLCwwZEhMPFB0aHx4dGhwcICQuJyAiLCMcHCg3KSwwMTQ0NB8nOT04MjwuMzQy/9sAQwEJCQkMCwwYDQ0YMiEcITIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIy/8QAHwAAAQUBAQEBAQEAAAAAAAAAAAECAwQFBgcICQoL/8QAtRAAAgEDAwIEAwUFBAQAAAF9AQIDAAQRBRIhMUEGE1FhByJxFDKBkaEII0KxwRVS0fAkM2JyggkKFhcYGRolJicoKSo0NTY3ODk6Q0RFRkdISUpTVFVWV1hZWmNkZWZnaGlqc3R1dnd4eXqDhIWGh4iJipKTlJWWl5iZmqKjpKWmp6ipqrKztLW2t7i5usLDxMXGx8jJytLT1NXW19jZ2uHi4+Tl5ufo6erx8vP09fb3+Pn6/8QAHwEAAwEBAQEBAQEBAQAAAAAAAAECAwQFBgcICQoL/8QAtREAAgECBAQDBAcFBAQAAQJ3AAECAxEEBSExBhJBUQdhcRMiMoEIFEKRobHBCSMzUvAVYnLRChYkNOEl8RcYGRomJygpKjU2Nzg5OkNERUZHSElKU1RVVldYWVpjZGVmZ2hpanN0dXZ3eHl6goOEhYaHiImKkpOUlZaXmJmaoqOkpaanqKmqsrO0tba3uLm6wsPExcbHyMnK0tPU1dbX2Nna4uPk5ebn6Onq8vP09fb3+Pn6/9oADAMBAAIRAxEAPwD5/oAKACgAoAKANKwsLxNRtWa0nVVmQkmMgAZHtXNWrU3TklJbPqe1l2XYyGMpSlSkkpR+y+68ixrlldzazcPHazOh24ZYyQflFZ4SrTjRScl9/mdvEGAxVXMak6dKTTtqk2vhXkYtdp8wFAG1Za5qM1/bxvcZR5VVhsXkE/SuKrhKMacml0fc+nwHEGY1cVSpzqXTkk9I7NryJ9Y1i/tdVmhhn2xrtwNin+EHuKzw2GpTpKUlr8+51Z3nePw2PqUaNS0VaysuyfVHPV6J8cFABQAUAFAH/9k=";
+
+        // Convert the Vec<u8> to base64 string for comparison
+        let actual_image_base64 = general_purpose::STANDARD.encode(&test_image);
+
+        assert_eq!(actual_image_base64, expected_image_base64);
+
+        // Clean up: remove the test image file
+        fs::remove_file(image_path).unwrap();
     }
 }