+++ src/main.rs
@@ -3,12 +3,11 @@ use std::io::Write;
 
 use anyhow::anyhow;
 use anyhow::Result;
+use aws_sdk_bedrockruntime::types::InferenceConfiguration;
 use bedrust::utils;
 
-use bedrust::ask_bedrock;
 use bedrust::configure_aws;
 use bedrust::utils::prompt_for_model_selection;
-use bedrust::RunType;
 
 use bedrust::captioner::caption_image;
 use bedrust::captioner::list_files_in_path_by_extension;
@@ -20,6 +19,19 @@ use clap::Parser;
 
 use bedrust::code::code_chat;
 use bedrust::constants;
+use bedrust::models::converse_stream::call_converse_stream;
+use bedrust::models::{check_model_features, ModelFeatures};
+
+// TODO:
+// So far I've implemented the converse API for general purpose chat and the code chat.
+// What I need to do is:
+// - Support Images and captioning - [DONE] ✅
+//  - Check if model supports images before attempting to run - [DONE] ✅
+// - Store the default inference parameters in the config file - [DONE] ✅
+// - Figure out feature support matrix for the Converse API and the models. - [DONE] ✅
+// - Remove nom v3.2.1 - [DONE] ✅
+// - Remove unwanted commented out code - [DONE] ✅
+// - Make sure everything works after ripping out the old code - [DONE] ✅
 
 #[tokio::main]
 async fn main() -> Result<()> {
@@ -81,84 +93,107 @@ async fn main() -> Result<()> {
     let bedrock_client = aws_sdk_bedrock::Client::new(&config);
 
     //let question = "Which songs are listed in the youtube video 'evolution of dance'?";
-    let model_id = arguments
-        .model_id
-        .or(bedrust_config.default_model);
+    let model_id = arguments.model_id.or(bedrust_config.default_model);
     let model_id = match model_id {
         Some(model_id) => model_id,
-        None => prompt_for_model_selection()?
-    }.to_str();
+        None => prompt_for_model_selection()?,
+    }
+    .to_str();
+
+    // === DEFAULT INFERENCE PARAMETERS ===
+    let inference_parameters = InferenceConfiguration::builder()
+        .max_tokens(bedrust_config.inference_params.max_tokens)
+        .top_p(bedrust_config.inference_params.top_p)
+        .temperature(bedrust_config.inference_params.temperature)
+        .build();
 
     // if we enabled captioning of images
     if arguments.caption.is_some() {
-        // FIX: This should be a function
-        println!("----------------------------------------");
-        println!("🖼️ | Image captioner running.");
-        let path = arguments
-            .caption
-            .ok_or_else(|| anyhow!("No path specified"))?;
-        println!("⌛ | Processing images in: {:?}", &path);
-        let files = list_files_in_path_by_extension(path, bedrust_config.supported_images)?;
-        println!("🔎 | Found {:?} images in path.", &files.len());
+        match check_model_features(model_id, &bedrock_client, ModelFeatures::Images).await {
+            Ok(b) => {
+                match b {
+                    true => {
+                        // FIX: This should be a function
+                        println!("----------------------------------------");
+                        println!("🖼️ | Image captioner running.");
+                        let path = arguments
+                            .caption
+                            .ok_or_else(|| anyhow!("No path specified"))?;
+                        println!("⌛ | Processing images in: {:?}", &path);
+                        let files =
+                            list_files_in_path_by_extension(path, bedrust_config.supported_images)?;
+                        println!("🔎 | Found {:?} images in path.", &files.len());
 
-        let mut images: Vec<Image> = Vec::new();
-        for file in &files {
-            images.push(Image::new(file)?);
-        }
+                        let mut images: Vec<Image> = Vec::new();
+                        for file in &files {
+                            images.push(Image::new(file)?);
+                        }
 
-        // TODO: Broken when used with Claude 3.5
-        caption_image(
-            &mut images,
-            model_id,
-            &bedrust_config.caption_prompt,
-            &bedrock_runtime_client,
-            &bedrock_client,
-        )
-        .await?;
+                        caption_image(
+                            &mut images,
+                            model_id,
+                            &bedrust_config.caption_prompt,
+                            &bedrock_runtime_client,
+                            &bedrock_client,
+                        )
+                        .await?;
 
-        // NOTE: This is parsing the `-x` argument and then writing or not, an XML file
-        // Thanks StellyUK <3
+                        // NOTE: This is parsing the `-x` argument and then writing or not, an XML file
+                        // Thanks StellyUK <3
 
-        // FIX: This whole if else statement does not look nice.
-        // i feel it can be better. As doing the whole logic
-        // behind an expression seems ... weird
-        let outfile = if arguments.xml {
-            let outfile = "captions.xml";
-            write_captions(images, OutputFormat::Xml, outfile)?;
-            outfile
-        } else {
-            let outfile = "captions.json";
-            write_captions(images, OutputFormat::Json, outfile)?;
-            outfile
+                        // FIX: This whole if else statement does not look nice.
+                        // i feel it can be better. As doing the whole logic
+                        // behind an expression seems ... weird
+                        let outfile = if arguments.xml {
+                            let outfile = "captions.xml";
+                            write_captions(images, OutputFormat::Xml, outfile)?;
+                            outfile
+                        } else {
+                            let outfile = "captions.json";
+                            write_captions(images, OutputFormat::Json, outfile)?;
+                            outfile
+                        };
+                        println!(
+                            "✅ | Captioning complete, find the generated captions in `{}`",
+                            outfile
+                        );
+                        println!("----------------------------------------");
+                    }
+                    false => {
+                        eprintln!("The current model selected does not support Images. Please consider using one that does.")
+                    }
+                }
+            }
+            Err(e) => eprintln!("Unable to determine model features: {}", e),
         };
-        println!(
-            "✅ | Captioning complete, find the generated captions in `{}`",
-            outfile
-        );
-        println!("----------------------------------------");
     } else {
         // default run
         utils::hello_header("Bedrust")?;
 
         // BETA: SOURCE READING
-        let mut conversation_history =  if arguments.source.is_some() {
+        let mut conversation_history = if arguments.source.is_some() {
             println!("----------------------------------------");
             print_warning("⚠ THIS IS A BETA FEATURE ⚠");
             println!("----------------------------------------");
             println!("💾 | Ooh, it Seems we are talking about code today!");
-            println!("💾 | I was given this dir to review: {:?}", arguments.source.clone().unwrap().into_os_string());
+            println!(
+                "💾 | I was given this dir to review: {:?}",
+                arguments.source.clone().unwrap().into_os_string()
+            );
             println!("----------------------------------------");
             let mut convo = String::new();
             convo.push_str(constants::CODE_CHAT_PROMPT);
 
-            let code = code_chat(arguments.source.clone().unwrap(), &bedrock_runtime_client).await?;
+            let code =
+                code_chat(arguments.source.clone().unwrap(), &bedrock_runtime_client).await?;
             println!("----------------------------------------");
             print_warning("⚠ THIS IS A BETA FEATURE ⚠");
 
             // Return this conversation
             convo.push_str(code.as_str());
             convo
-        } else { // We are not looking at code
+        } else {
+            // We are not looking at code
             // Just return an empty string
             String::new()
         };
@@ -199,17 +234,15 @@ async fn main() -> Result<()> {
             println!("----------------------------------------");
             println!("☎️  | Calling Model: {}", &model_id);
             println!("----------------------------------------");
-            let response = ask_bedrock(
-                //&question.to_string(),
-                &conversation_history.to_string(),
-                None,
-                model_id,
-                RunType::Standard,
+
+            let streamresp = call_converse_stream(
                 &bedrock_runtime_client,
-                &bedrock_client,
+                model_id.to_string(),
+                &conversation_history.to_string(),
+                inference_parameters.clone(),
             )
             .await?;
-            conversation_history.push_str(&response);
+            conversation_history.push_str(&streamresp);
             conversation_history.push('\n');
         }
     }